10.4.Handling Events (like onClick) ! {Moving the Handler to the Class}
Moving the Handler to the Class

Putting all this logic inline is fine... but it can become hard to read. So, instead, I like to make the handler a property on my class. Start by adding a new method: handleRowClick that will accept the repLogId that was just clicked and also the event object itself... just to show that we can we pass this:

Next, steal the state-setting code and paste it here, but with highlightedRowId set to repLogId. And... we should probably close the method so Webpack isn't so mad at me!

Below, call it: this.handleRowClick() with repLog.id and event.

I like it! Let's make sure we didn't bork our cool app: back to the browser! Refresh! Yea! It still works!

This is the power of React! It doesn't care how many different things in your UI need to change when some state changes, it takes care of everything.

And now, it's time to talk about organization. RepLogApp is big, and when things get too big, they get confusing. Let's move some code into a new child component.