14.1.Prop Validation: PropTypes{yarn add prop-types --dev}
Look at RepLogList: it uses 2 props: highlightedRowId and onRowClick. But, what guarantees that RepLogList is actually passed these props? What prevents us from forgetting to pass those two exact props from RepLogs?

The answer is... absolutely nothing! In PHP, when you instantiate an object, if that object has a constructor with required arguments, we are forced to pass those arguments. But, with React components, it's the wild west! There is nothing like that. There is simply no guarantee that we are passed any of these props. There's also no way for a component to easily document, or advertise what props it needs.

This is an especially big problem with onRowClick, because, if we forget to pass this prop, our whole app will break when someone clicks on a row.

Introducing PropTypes

To fix this, React leverages something called "prop types", which uses an external library. Find your terminal and install it:

yarn add prop-types --dev