14.2.Prop Validation: PropTypes{PropTypes Give us Warnings}
With prop types, we can tell React exactly which props a component needs, and even what they should look like. At the bottom of the component, add RepLogList.propTypes = and pass it an object. This is where you describe all the different props this component might have. But first, back on top, import PropTypes from prop-types. And... back down below, add highlightedRowId set to PropTypes.any. Then, onRowClick set to PropTypes.func.

For highlightedRowId, we could have used PropTypes.number, because, it is a number right now. But later, we're going to refactor our ids to be uuid's, which are a string.

PropTypes Give us Warnings

Let's see this in action! Go back to RepLogs and, instead of passing in a function, be devious: pass a string!

Check it out: move back to your browser! An error! Oh, and, notice: the page refreshed automatically before we got here. That's thanks to the Encore dev-server we're running: when we save a file, our browser automatically refreshes, which, is kinda nice.

Anyways, we see:

  Invalid prop onRowClick of type string supplied to RepLogList, expected function.

This is what PropTypes give you: clear & early warnings when we mess up. I mess up a lot!