14.4.Prop Validation: PropTypes{PhpStorm ❤️'s PropTypes! }
PhpStorm ❤️'s PropTypes!

I love it! Enough fun: let's re-add the prop. Woh! We're now getting auto-complete! Yeaaa! PhpStorm reads the propTypes and uses them to help us out.

And, in RepLogList, before, we had big red warnings when we referenced our props. That came from ESLint: it was telling us that we forgot to add the prop types. Now, everyone is happy.

Adding propTypes Everywhere Else

Because propTypes are amazing, let's add them everywhere else, like in RepLogApp. Here, we're relying on a withHeart prop. And, there is the warning I was just talking about:

  Missing props validation