14.5.Prop Validation: PropTypes{Adding propTypes Everywhere Else }
Adding propTypes Everywhere Else

Because propTypes are amazing, let's add them everywhere else, like in RepLogApp. Here, we're relying on a withHeart prop. And, there is the warning I was just talking about:

  Missing props validation

You guys know the drill! First, import PropTypes from prop-types. Then, at the bottom, RepLogApp.propTypes = an object with withHeart set to PropTypes.bool. The prop isn't really required, so I'll leave it optional.

The last place we need propTypes is in RepLogs: we depend on 3. Go copy the import statement from RepLogApp, and paste! At the bottom, add the RepLogs.propTypes = part.

The 3 props are: withHeart, highlightedRowId and onRowClick. Steal withHeart from RepLogApp and paste. Then, more stealing! Get the other two from RepLogList and put those here too.

Hmm, this shows off another common thing in React. Frequently, you'll pass props into one component, just so that it can pass them into another component. For example, in RepLogApp, we pass 3 props. But, two of them aren't event used in RepLogs! We just pass them straight to RepLogList!

This "props passing" can be kind of annoying. But, it's not necessarily a sign of bad design. It's just part of using React. There are ways to organize our code to help this, but many are more advanced. The point is: this is ok.

Next, let's make a small production optimization with prop types.