15.3.Removing propTypes on Production{}
Oh wait, actually, I made a mistake! This totally won't work! That's because we can't rely on Babel to know whether or not we're creating our production build. Instead, use if Encore.isProduction(). Then, inside, add the plugin with babelConfig.plugins.push(), copy the plugin name, and paste!

Remove the stuff below. This is simpler anyways: if we're building for production, add this handy plugin.

We're not going to build for production right now, but to make sure we didn't break anything, go back to the terminal that runs encore, press Ctrl+C to stop it, then restart:

yarn run encore dev-server

And... no errors! Later, when we execute yarn run encore production, the prop types won't be there.