16.1.Moving the Rep Logs to State{Why state Lives up High}
Communication always flows down in React: data lives in one component and is passed down to its children as props. Actually, both data and callbacks are passed from parent to child: child components use callbacks to communicate back up to the parent when something happens. For example, RepLogs passes an onRowClick to RepLogList. It uses that to tell its parent when that interaction occurs.

So, parents pass data to their children. But, parents do not, ask children for information. Well, it's technically possible, but it's not the normal flow.

For example, RepLogApp passes the highlightedRowId to RepLogs. But, RepLogApp does not ever ask RepLogs to give it any data that lives inside RepLogs. Information only flows down.

Why state Lives up High

For that reason, in general, we need the state of our application to live as high up the component hierarchy as possible. Why? Because we can pass a piece of state down to all of the components that need to use it. When that state changes, that change naturally flows down and everything updates beautifully.

But, imagine if a piece of state lived in a child component, but we wanted to use it in the render() method of a parent. Well, that just won't work! The parent can't ask the child for that data: information does not flow up.

This is the reason why we will put all of our state in the top level component: RepLogApp. Again, this is not an absolute rule, but it's a great rule to follow for now. We'll talk later about when it's ok to move state lower, into a child component.