16.2.Moving the Rep Logs to State{Moving RepLogs to state; Passing the RepLogs State Down}
Moving RepLogs to state

Anyways, the most important piece of data in our app is the rep logs themselves. And, these will need to change dynamically as the user adds new rep logs and deletes old ones. That means, rep logs need to be stored as state.

To get the static version of our app up and running, we just hardcoded these inside RepLogList. Time to move this to state! Copy the dummy rep log data and go to RepLogApp. Whenever we have new state, we need to initialize it in the constructor. Add a new repLogs key to this array and paste!

Yea, eventually the repLogs state will start empty, and we'll then populate it by making an AJAX call to the server for the existing rep logs. But, until then, the dummy data makes building things easier.

Passing the RepLogs State Down

The new state lives in the top-level component. But... we need to use it down in RepLogList. No problem! We just need to pass this down our tree. Fetch the repLogs out of state, then pass this as a prop to RepLogs.

In RepLogs, before using the new prop, head down to the bottom: we want to define all props in propTypes. Add repLogs set to PropTypes.array.isRequired.

Copy that, because, RepLogList will receive the same prop.

Ok! We are passing the repLogs prop to the RepLogs component. At the top of render(), read repLogs out of props. And then, do the prop-passing dance: send this straight into RepLogList.

Finally, in that component, get repLogs out of props and... delete the hardcoded stuff.

This is sweet! Move back to your browser and refresh! Hey hey! It's not broken! Check out the React dev tools, and look at the top RepLogApp component. Yep! You can see the repLogs state. Now... mess with it! Change the reps from 25 to 50.... boom! The UI on the child component updates instantly!

But, look back at RepLogApp, it has two pieces of state & one prop. And... it's passing all of that into its child as props. With a trick, we can be lazier, and do this automatically.