17.1.Smart Components & Spread Attributes{The Attribute Spread Operator}
RepLogApp is a "smart", or "container" component: it holds state & logic... but no markup. Instead, smart components wrap themselves around a dumb component - like RepLogs - and that component renders all the elements.

Thanks to this pattern... a funny thing happens in smart components: you pretty much always want to pass all of your state and props into your child so that it can actually use them to render!

But, this can become tedious: each new piece of state needs to be initialized, destructured into a variable, then passed as a prop... with the same name. Lame! Let's use a shortcut.

The Attribute Spread Operator

Delete all of the variables we destructured from state and props. Then, delete all of the props we're passing to RepLogs, except for the callback. Instead, here's the awesome part, use the spread operator: {...this.props} and {...this.state}.

That's it! Every prop and state is now being passed to the child as a prop. The only things we need to pass by hand are any callbacks we need.

Move over and refresh. No problems. Our smart component is getting simpler and simpler: it's all logic.