17.2.Smart Components & Spread Attributes{Calculating the Total Weight Lifted}
Calculating the Total Weight Lifted

While we're here, it's time to finish one of our TODO's: fill in the total weight lifted column. This is great! I usually end up forgetting about my TODO's until I accidentally find them years later.

The value in this column will need to change whenever the repLogs state changes. But... the total weight should not be stored in state! Why? Simple! We can already calculate it by adding up all of the total weight values for the rep logs. No need to introduce more state: that would be duplication.

Go to RepLogs: there's the TODO! And here's the plan: loop over the repLogs and add up the weights. So, we need just a little bit of logic. If this component were a class, I'd probably add a new method to the class and put that logic there. But, darn! It's just a function! No problem: just go above the function and add... another function! Call it calculateTotalWeightLifted() with a repLogs argument: we will need to pass that in.

To do the calculation, I'll paste in some boooooring code: it loops over the repLogs, adds up the totalWeightLifted for each and... returns.

Copy the function name, move down to the TODO, and call it! {calculateTotalWeightLifted()} passing it repLogs. The repLogs live inside props, but we already destructured that into a variable.

Moment of truth: refresh! Boom! We have a total! Let's mess with the state, like change this to 200. Yes! It updates! The state change on RepLogApp causes both RepLogApp and RepLogList to re-render. When that happens, our code uses the new weights to calculate the new total. It's all very awesome.