17.3.Smart Components & Spread Attributes{Super-Fancy (Confusing?) ES6 Syntax}
Super-Fancy (Confusing?) ES6 Syntax

This works fine. But, I'm going to write a new function that does this same calculation... but with some crazy, fancy syntax. Check this out:

const calculateTotalWeightFancier = then repLogs => repLogs.reduce(). Pass this another arrow function with two arguments: total and log. That callback will return total + log.totalWeightLifted. Start the reduce function with a 0 value.

Phew! Before we understand this madness, copy the new function name, move down, and paste! Find your browser - ah, the page is already reloading... and... it works!

This fancier function doesn't contain anything new. But wow, even for me, this is hard to understand. So, why are we doing this? Because, in the React world, you will see syntax like this. And I want you to be at least comfortable reading it.

Let's walk through it.

This creates a variable that is set to a function that accepts one argument: repLogs. Because the function doesn't have curly braces, it means the function returns the result of repLogs.reduce(). The reduce() function - which you may not be familiar with - itself accepts a callback function with two arguments. Once again, because that function doesn't have curly braces, it means that it returns total + log.totalWeightLifted.

If this makes your head spin, me too! Honestly, to me, this looks much more complex than the original. But, when you see things like this in blog posts or documentation, just break it down piece by piece: it's just boring code, dressed up in a different style. And if you like this syntax, cool! Go nuts.