19.3.New Component to Hold our Form{Creating RepLogCreator}
Go grab handleFormSubmit from RepLogs - and, by the way - check out how small this component is looking - then, inside RepLogCreator, paste.

The last missing piece is the onNewItemSubmit() callback: this is passed from RepLogApp to RepLogs. And now, we need to pass it once again to RepLogCreator. Because we need a new prop, define it first at the bottom in propTypes: RepLogCreator.propTypes = an object and... go steal this code from RepLogs.

Excellent! Now that we are requiring this prop, head back up to render() and destructure it: const { onNewItemSubmit } = props.

Cool! Finally, in RepLogs, nice! PhpStorm is already telling us that we're missing a required prop: pass this onNewItemSubmit={onNewItemSubmit}.

And... we're done! Probably... Let's go find out: refresh. The form renders... we can select something and... it does print. Awesome!